api/pyproject.toml
[project]
name = "cache-layer-api"
version = "0.1.0"
description = "Cache Layer API - Intelligent response caching for AI applications"
readme = "../ARCHITECTURE.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "mangum>=0.18.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "boto3>=1.35.0",
    "ulid-py>=1.1.0",
    "httpx>=0.28.0",
    "boldsci-auth",
    "structlog>=24.0.0",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-instrumentation-fastapi>=0.41b0",
    "opentelemetry-instrumentation-botocore>=0.41b0",
    "opentelemetry-instrumentation-httpx>=0.41b0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "moto[dynamodb]>=5.0.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
    "mypy>=1.14.0",
    "boto3-stubs[dynamodb]>=1.35.0",
]

[tool.uv.sources]
boldsci-auth = { git = "https://github.com/BOLDSCIENCE/platform-block-auth.git", subdirectory = "sdk" }

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.ruff.format]
quote-style = "double"

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
